renamed `TFPipeline` to `AtomsiticDataset`. Moved DS initialization to submodule. DS no longer initializes with batch size. Moved dict to DS conversion to function
